shopweb/requirements.txt
pandas
django==2.1.5
googlemaps
sqlalchemy
pymysql
matplotlib
requests
bs4
lxml
opencc-python-reimplemented
jieba
sklearn 
joblib
numpy
bs4
textrank4zh
snownlp
shortcuts
transformers==2.1.1
wordcloud
elasticsearch
argparse
djangorestframework
scrapy
Twisted
channels==2.1.7
channels_redis
aiowebsocket
asset
websocket
websocket-client 
tornado
djangorestframework==3.9.4
mysqlclient-1.4.6-cp37-cp37m-win_amd64.whl